tdf#58511 sw: Fix source fontwork object not removed after being cut
In the `SwWrtShell::DelRight` class method where the cut operation logic
will call, the `GetSelectionType()` call returned `0x10200` for this
object, which is a bitwise OR-ed value of `SelectionType::FontWork` and
`SelectionType::DrawObject` taht is assigned to the `nSelection`
variable. The variable won't match any cases in the subsequent `switch`
statement where the object removal operation will be done as all the
cases only match a single `SelectionType` value.

This patch adds a `if` statement to set the `nSelection` variable
to `SelectionType::DrawObject` when the object is a FontWork, making
it enter the intended logic to handle such an object's deletion.

Signed-off-by: Buo-ren Lin (OSSII) <buoren.lin@ossii.com.tw>
Reviewed-by: Franklin Weng <franklin@goodhorse.idv.tw>
Change-Id: I8074a98be9d10fc26f8d498abfe94b011cc97eb0